Refactoring AlarmService
